module_5/url_shortener/pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = "URL Shortener Microservice"
authors = ["Bright <brightcollin1619@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.1"
djangorestframework = "^3.16.1"
redis = "^7.1.0"
python-decouple = "^3.8"
drf-spectacular = "^0.29.0"
gunicorn = "^24.1.1"
psycopg2-binary = "^2.9.11"
django-redis = "^6.0.0"
whitenoise = "^6.11.0"
djangorestframework-simplejwt = "^5.5.0"
celery = "^5.3.6"
httpx = "^0.28.1"
tenacity = "^9.1.4"
beautifulsoup4 = "^4.14.3"
django-cors-headers = "^4.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
isort = "^5.13"
pytest = "^8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
